requirements.txt
Django==2.1.7
django-oidc-provider==0.7.0
ipython==7.2.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
